+++ esmecata/precomputed/create_database.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2024 Arnaud Belcour - Univ. Grenoble Alpes, Inria, Microcosme
+# Copyright (C) 2024-2025 Arnaud Belcour - Univ. Grenoble Alpes, Inria, Microcosme
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
 # the Free Software Foundation, either version 3 of the License, or
@@ -17,12 +17,17 @@ import shutil
 import pandas as pd
 import json
 import csv
+import logging
 import zipfile
 
 from esmecata.core.eggnog import get_proteomes_tax_id_name
 
 from multiprocessing import Pool
 from collections import Counter
+
+from ete3 import NCBITaxa
+
+logger = logging.getLogger(__name__)
 
 
 def get_proteomes_tax_id_name(proteomes_tax_id_file_path):
@@ -135,106 +140,17 @@ def copy_file(annotation_file, proteomes_taxa_names, computed_threshold_folder, 
         shutil.copyfile(consensus_sequence_file_path, taxon_consensus_sequence_file_path)
 
 
-def concat_tsv_file(input_tsv_file_path, tsv_file_concat_path, consensus_sequence_folder, annotation_reference_folder):
-    """ Concat tsv file for database by checking that fasta and tsv files exist.
-
-    Args:
-        input_tsv_file_path (str): path to input tsv file.
-        tsv_file_concat_path (str): path to tsv file containing concatenation.
-        consensus_sequence_folder (str): path to consensus folder.
-        annotation_reference_folder (str): path to annotation reference folder.
-    """
-    if not os.path.exists(tsv_file_concat_path):
-        df_proteomes_tax_id_file_path = pd.read_csv(input_tsv_file_path, sep='\t')
-        # Check that corresponding files exist.
-        existing_files = [os.path.exists(consensus_sequence_folder, tax_id_name+'.faa') and os.path.exists(annotation_reference_folder, tax_id_name+'.tsv')
-            for tax_id_name in df_proteomes_tax_id_file_path['tax_id_name']]
-        df_proteomes_tax_id_file_path = df_proteomes_tax_id_file_path[existing_files]
-        df_proteomes_tax_id_file_path.to_csv(tsv_file_concat_path, sep='\t', index=False)
-    else:
-        df_proteomes_tax_id_file_path = pd.read_csv(input_tsv_file_path, sep='\t')
-        # Check that corresponding files exist.
-        existing_files = [os.path.exists(consensus_sequence_folder, tax_id_name, tax_id_name+'.faa') and os.path.exists(annotation_reference_folder, tax_id_name, tax_id_name+'.tsv')
-            for tax_id_name in df_proteomes_tax_id_file_path['tax_id_name']]
-        df_proteomes_tax_id_file_path = df_proteomes_tax_id_file_path[existing_files]
-        df_tsv_file_concat = pd.read_csv(tsv_file_concat_path, sep='\t')
-        df_concat = pd.concat([df_proteomes_tax_id_file_path, df_tsv_file_concat])
-        df_concat.to_csv(tsv_file_concat_path, sep='\t', index=False)
-
-
-def create_database(esmecata_proteomes_folder, esmecata_clustering_folder, esmecata_annotation_folder, output_database_folder, nb_core=1):
+def create_database_from_run(esmecata_proteomes_folder, esmecata_clustering_folder, esmecata_annotation_folder, output_folder, nb_core=1):
     """ Create esmecata database from a run of esmecata.
 
     Args:
         esmecata_proteomes_folder (str): path to esmecata proteomes folder.
         esmecata_clustering_folder (str): path to esmecata clustering folder.
         esmecata_annotation_folder (str): path to esmecata anntoation folder.
-        output_database_folder (str): path to output folder containing zip database of esmecata.
+        output_folder (str): path to output folder.
         nb_core (int): number of core to use when creating database.
     """
-    if not os.path.exists(output_database_folder):
-        os.mkdir(output_database_folder)
-
-    consensus_sequence_folder = os.path.join(esmecata_clustering_folder, 'reference_proteins_consensus_fasta')
-    computed_threshold_folder = os.path.join(esmecata_clustering_folder, 'computed_threshold')
-
-    annotation_reference_folder = os.path.join(esmecata_annotation_folder, 'annotation_reference')
-
-    # Create/merge proteome_tax_id file for each taxon level.
-    proteomes_tax_id_file_path = os.path.join(esmecata_clustering_folder, 'proteome_tax_id.tsv')
-    proteomes_tax_id_file_database_path = os.path.join(output_database_folder, 'proteome_tax_id.tsv')
-    concat_tsv_file(proteomes_tax_id_file_path, proteomes_tax_id_file_database_path, consensus_sequence_folder, annotation_reference_folder)
-
-    # Create/merge stat_proteome file for each taxon level.
-    stat_number_proteome_file_path = os.path.join(esmecata_proteomes_folder, 'stat_number_proteome.tsv')
-    stat_number_proteome_database_path = os.path.join(output_database_folder, 'stat_number_proteome.tsv')
-    concat_tsv_file(stat_number_proteome_file_path, stat_number_proteome_database_path)
-
-    # Create/merge stat_number_clustering file for each taxon level.
-    stat_number_clustering_file_path = os.path.join(esmecata_clustering_folder, 'stat_number_clustering.tsv')
-    stat_number_clustering_database_path = os.path.join(output_database_folder, 'stat_number_clustering.tsv')
-    concat_tsv_file(stat_number_clustering_file_path, stat_number_clustering_database_path)
-
-    # Create/merge stat_proteome file for each taxon level.
-    stat_number_annotation_file_path = os.path.join(esmecata_annotation_folder, 'stat_number_annotation.tsv')
-    stat_number_annotation_database_path = os.path.join(output_database_folder, 'stat_number_annotation.tsv')
-    concat_tsv_file(stat_number_annotation_file_path, stat_number_annotation_database_path)
-
-    proteomes_taxa_names = get_proteomes_tax_id_name(proteomes_tax_id_file_path)
-
-    # Use multiprocessing to copy fasta and annotation files.
-    database_copy_pool = Pool(nb_core)
-
-    similar_tax_id_names = []
-    for observation_name in proteomes_taxa_names:
-        similar_tax_id_names.append(proteomes_taxa_names[observation_name])
-    similar_tax_id_names = Counter(similar_tax_id_names)
-
-    multiprocessing_data = []
-    for annotation_file in os.listdir(annotation_reference_folder):
-        observation_name = os.path.splitext(annotation_file)[0]
-        taxon_id_name = proteomes_taxa_names[observation_name]
-        # If more than 1 observation name associated with a similar tax_id_name, do not use multiprocessing as it leads error.
-        if similar_tax_id_names[taxon_id_name] > 1:
-            copy_file(annotation_file, proteomes_taxa_names, computed_threshold_folder, annotation_reference_folder, consensus_sequence_folder, output_database_folder)
-        else:
-            multiprocessing_data.append([annotation_file, proteomes_taxa_names, computed_threshold_folder, annotation_reference_folder, consensus_sequence_folder, output_database_folder])
-    database_copy_pool.starmap(copy_file, multiprocessing_data)
-
-    database_copy_pool.close()
-    database_copy_pool.join()
-
-
-def create_database_from_run(esmecata_proteomes_folder, esmecata_clustering_folder, esmecata_annotation_folder, output_database_folder, nb_core=1):
-    """ Create esmecata database from a run of esmecata.
-
-    Args:
-        esmecata_proteomes_folder (str): path to esmecata proteomes folder.
-        esmecata_clustering_folder (str): path to esmecata clustering folder.
-        esmecata_annotation_folder (str): path to esmecata anntoation folder.
-        output_database_folder (str): path to output folder containing zip database of esmecata.
-        nb_core (int): number of core to use when creating database.
-    """
+    output_database_folder = os.path.join(output_folder, 'database_folder')
     if not os.path.exists(output_database_folder):
         os.mkdir(output_database_folder)
 
@@ -242,6 +158,19 @@ def create_database_from_run(esmecata_proteomes_folder, esmecata_clustering_fold
     proteomes_tax_id_file_path = os.path.join(esmecata_clustering_folder, 'proteome_tax_id.tsv')
     df_proteomes_tax_id = pd.read_csv(proteomes_tax_id_file_path, sep='\t')
     df_proteomes_tax_id.set_index('observation_name', inplace=True)
+    obs_name_proteomes = df_proteomes_tax_id['proteome'].to_dict()
+
+    # Get dict mapping tax IDs and observation names.
+    obs_name_tax_ids = df_proteomes_tax_id['tax_id'].to_dict()
+    tax_id_obs_names = {}
+    for observation_name in obs_name_tax_ids:
+        tax_id = obs_name_tax_ids[observation_name]
+        if tax_id not in tax_id_obs_names:
+            tax_id_obs_names[tax_id] = [observation_name]
+        else:
+            tax_id_obs_names[tax_id].append(observation_name)
+
+    esmecata_proteomes_downloaded_folder = os.path.join(esmecata_proteomes_folder, 'proteomes')
 
     # Create/merge stat_proteome file for each taxon level.
     stat_number_proteome_file_path = os.path.join(esmecata_proteomes_folder, 'stat_number_proteome.tsv')
@@ -252,6 +181,10 @@ def create_database_from_run(esmecata_proteomes_folder, esmecata_clustering_fold
     stat_number_clustering_file_path = os.path.join(esmecata_clustering_folder, 'stat_number_clustering.tsv')
     df_stat_number_clustering = pd.read_csv(stat_number_clustering_file_path, sep='\t')
     df_stat_number_clustering.set_index('observation_name', inplace=True)
+
+    df_stat_join = df_proteomes_tax_id.join(df_stat_number_clustering)
+    df_stat_join.set_index('tax_id', inplace=True)
+    tax_id_protein_clusters = df_stat_join['Number_protein_clusters_kept'].to_dict()
 
     # Create/merge stat_proteome file for each taxon level.
     stat_number_annotation_file_path = os.path.join(esmecata_annotation_folder, 'stat_number_annotation.tsv')
@@ -265,6 +198,49 @@ def create_database_from_run(esmecata_proteomes_folder, esmecata_clustering_fold
     reference_proteins_folder = os.path.join(esmecata_clustering_folder, 'reference_proteins')
 
     annotation_reference_folder = os.path.join(esmecata_annotation_folder, 'annotation_reference')
+
+    ncbi = NCBITaxa()
+    database_tree = ncbi.get_topology([org_tax_id for org_tax_id in df_proteomes_tax_id['tax_id']])
+    tax_id_issues = {}
+    # Parse all descendant of root to search for children tax id with lower protein than parents.
+    # To identify taxon with potential errors.
+    for descendant in database_tree.iter_descendants():
+        descendant_childrens = descendant.children
+        parent_tax_id = int(descendant.name)
+        if parent_tax_id in tax_id_protein_clusters:
+            parent_protein_nb = tax_id_protein_clusters[parent_tax_id]
+        else:
+            parent_protein_nb = None
+        if descendant_childrens != []:
+            for children in descendant.children:
+                children_tax_id = int(children.name)
+                if children_tax_id in tax_id_protein_clusters:
+                    children_protein_nb = tax_id_protein_clusters[children_tax_id]
+                else:
+                    children_protein_nb = None
+                if children_protein_nb is not None and parent_protein_nb is not None:
+                    # Compute percent of proteins in children compare to parent
+                    percent_protein_nb = (children_protein_nb / parent_protein_nb) * 100
+                    # If children has less than 20% of proteins compared to the parent, add them to the issues.
+                    if percent_protein_nb < 20:
+                        tax_id_issues[children_tax_id] = (children_protein_nb, parent_tax_id, parent_protein_nb)
+
+    predictions_with_issues = []
+    issues_tax_id_folder = os.path.join(output_folder, 'issues_tax_id')
+    if not os.path.exists(issues_tax_id_folder):
+        os.mkdir(issues_tax_id_folder)
+    for tax_id in tax_id_issues:
+        children_protein_nb = tax_id_issues[tax_id][0]
+        parent_tax_id = tax_id_issues[tax_id][1]
+        parent_protein_nb = tax_id_issues[tax_id][2]
+        for observation_name in tax_id_obs_names[tax_id]:
+            observation_name_issue_folder = os.path.join(issues_tax_id_folder, observation_name)
+            if not os.path.exists(observation_name_issue_folder):
+                os.mkdir(observation_name_issue_folder)
+            for proteome in obs_name_proteomes[observation_name].split(','):
+                shutil.copyfile(os.path.join(esmecata_proteomes_downloaded_folder, proteome+'.faa.gz'), os.path.join(observation_name_issue_folder, proteome+'.faa.gz'))
+            logger.info(f'|EsMeCaTa|create_db| {observation_name} with less than 20% of protein clusters ({children_protein_nb}) compared to parent ({parent_tax_id}) protein clusters ({parent_protein_nb}).')
+            predictions_with_issues.append(observation_name)
 
     # Use multiprocessing to copy fasta and annotation files.
     database_copy_pool = Pool(nb_core)
@@ -282,20 +258,24 @@ def create_database_from_run(esmecata_proteomes_folder, esmecata_clustering_fold
     multiprocessing_data = []
     for annotation_file in os.listdir(annotation_reference_folder):
         observation_name = os.path.splitext(annotation_file)[0]
-        taxon_id_name = proteomes_taxa_names[observation_name]
-        reference_proteins_consensus_fasta_file = os.path.join(consensus_sequence_folder, taxon_id_name+'.faa')
-        if taxon_id_name not in already_processed_tax_id and os.path.exists(reference_proteins_consensus_fasta_file):
-            proteome_tax_id_data.append([taxon_id_name, *df_proteomes_tax_id.loc[observation_name].to_list()])
-            stat_number_proteome_data.append([taxon_id_name, *df_stat_number_proteome.loc[observation_name].to_list()])
-            stat_number_clustering_data.append([taxon_id_name, *df_stat_number_clustering.loc[observation_name].to_list()])
-            stat_number_annotation_data.append([taxon_id_name, *df_stat_number_annotation.loc[observation_name].to_list()])
-
-            already_processed_tax_id.append(taxon_id_name)
-        # If more than 1 observation name associated with a similar tax_id_name, do not use multiprocessing as it leads error.
-        if similar_tax_id_names[taxon_id_name] > 1:
-            copy_file(annotation_file, proteomes_taxa_names, computed_threshold_folder, annotation_reference_folder, consensus_sequence_folder, output_database_folder)
+        if observation_name in predictions_with_issues:
+            logger.info(f'|EsMeCaTa|create_db| {observation_name} will not be added to database as it contains 0 ECs and 0 GO Terms.')
         else:
-            multiprocessing_data.append([annotation_file, proteomes_taxa_names, computed_threshold_folder, annotation_reference_folder, consensus_sequence_folder, output_database_folder])
+            taxon_id_name = proteomes_taxa_names[observation_name]
+            reference_proteins_consensus_fasta_file = os.path.join(consensus_sequence_folder, taxon_id_name+'.faa')
+            if taxon_id_name not in already_processed_tax_id and os.path.exists(reference_proteins_consensus_fasta_file):
+                proteome_tax_id_data.append([taxon_id_name, *df_proteomes_tax_id.loc[observation_name].to_list()])
+                stat_number_proteome_data.append([taxon_id_name, *df_stat_number_proteome.loc[observation_name].to_list()])
+                stat_number_clustering_data.append([taxon_id_name, *df_stat_number_clustering.loc[observation_name].to_list()])
+                stat_number_annotation_data.append([taxon_id_name, *df_stat_number_annotation.loc[observation_name].to_list()])
+
+                already_processed_tax_id.append(taxon_id_name)
+            # If more than 1 observation name associated with a similar tax_id_name, do not use multiprocessing as it leads error.
+            if similar_tax_id_names[taxon_id_name] > 1:
+                copy_file(annotation_file, proteomes_taxa_names, computed_threshold_folder, annotation_reference_folder, consensus_sequence_folder, output_database_folder)
+            else:
+                multiprocessing_data.append([annotation_file, proteomes_taxa_names, computed_threshold_folder, annotation_reference_folder, consensus_sequence_folder, output_database_folder])
+
     database_copy_pool.starmap(copy_file, multiprocessing_data)
 
     database_copy_pool.close()
@@ -333,7 +313,7 @@ def create_database_from_esmecata_run(esmecata_proteomes_folder, esmecata_cluste
         os.mkdir(output_database_folder)
 
     create_json(esmecata_proteomes_folder, esmecata_clustering_folder, esmecata_annotation_folder, output_database_folder, database_version)
-    create_database_from_run(esmecata_proteomes_folder, esmecata_clustering_folder, esmecata_annotation_folder, output_database_folder, nb_core)
+    create_database_from_run(esmecata_proteomes_folder, esmecata_clustering_folder, esmecata_annotation_folder, output_folder, nb_core)
 
     compress_database_file = os.path.join(output_folder, 'esmecata_database')
     if not os.path.exists(compress_database_file):
@@ -357,38 +337,13 @@ def create_database_from_esmecata_workflow_run(esmecata_workflow_folder, output_
     create_database_from_esmecata_run(esmecata_proteomes_folder, esmecata_clustering_folder, esmecata_annotation_folder, output_folder, database_version, nb_core)
 
 
-def create_database_from_multiple_esmecata_runs(esmecata_proteomes_folders, esmecata_clustering_folders, esmecata_annotation_folders, output_folder, database_version, cpu_number=1):
-    """ Extract data from esmecata runs and create an esmecata database from these.
+def merge_db_files(list_db_files, output_folder):
+    """ From a list of zip files of precomputed database of esmecata, merged them in one database.
 
     Args:
-        esmecata_proteomes_folders (list): list of path to esmecata proteomes folder.
-        esmecata_clustering_folders (list): list of path to esmecata clustering folder.
-        esmecata_annotation_folders (list): list of path to esmecata anntoation folder.
+        list_db_files (list): list of path to precomputed zip file of esmecata.
         output_folder (str): path to output folder containing zip database of esmecata.
-        nb_core (int): number of core to use when creating database.
     """
-    if not os.path.exists(output_folder):
-        os.mkdir(output_folder)
-
-    output_database_folder = os.path.join(output_folder, 'database_folder')
-    if not os.path.exists(output_database_folder):
-        os.mkdir(output_database_folder)
-
-    for index, esmecata_proteomes_folder in enumerate(esmecata_proteomes_folders):
-        taxon_level = os.path.basename(esmecata_proteomes_folder).split('_')[0]
-        esmecata_clustering_folder = esmecata_clustering_folders[index]
-        esmecata_annotation_folder = esmecata_annotation_folders[index]
-        create_json(esmecata_proteomes_folder, esmecata_clustering_folder, esmecata_annotation_folder, output_database_folder, database_version, taxon_level)
-        create_database(esmecata_proteomes_folder, esmecata_clustering_folder, esmecata_annotation_folder, output_database_folder, cpu_number)
-
-    compress_database_file = os.path.join(output_folder, 'esmecata_database')
-    if not os.path.exists(compress_database_file):
-        os.mkdir(compress_database_file)
-    shutil.make_archive(compress_database_file, 'zip', output_database_folder)
-    shutil.rmtree(output_database_folder)
-
-
-def merge_db_files(list_db_files, output_folder):
     stat_number_proteome_df = pd.DataFrame(columns=['observation_name', 'Number_proteomes', 'Input_taxon_Name', 'Taxon_rank', 'EsMeCaTa_used_taxon', 'EsMeCaTa_used_rank', 'only_reference_proteome_used'])
     stat_number_clustering_df = pd.DataFrame(columns=['observation_name', 'Number_protein_clusters_panproteome', 'Number_protein_clusters_kept', 'Number_protein_clusters_coreproteome'])
     stat_number_annotation_df = pd.DataFrame(columns=['observation_name', 'Number_go_terms', 'Number_ecs'])